+++ application/network/src/main/kotlin/io/github/siyual_park/application/network/Application.kt
@@ -1,7 +1,7 @@
-package io.github.siyual_park
+package io.github.siyual_park.application.network
 
-import io.github.siyual_park.routes.getRootRoutes
-import io.github.siyual_park.routes.pingRoutes
+import io.github.siyual_park.application.network.routes.getRootRoutes
+import io.github.siyual_park.application.network.routes.pingRoutes
 import io.ktor.application.Application
 import io.ktor.application.install
 import io.ktor.routing.Routing
