+++ application/network/src/test/kotlin/io/github/siyual_park/application/network/integration/RootRoutesTests.kt
@@ -1,6 +1,6 @@
-package io.github.siyual_park.integration
+package io.github.siyual_park.application.network.integration
 
-import io.github.siyual_park.module
+import io.github.siyual_park.application.network.module
 import io.ktor.http.HttpMethod
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.testing.handleRequest
@@ -8,12 +8,12 @@ import io.ktor.server.testing.withTestApplication
 import org.junit.Test
 import kotlin.test.assertEquals
 
-class PingRoutesTests {
+class RootRoutesTests {
     @Test
     fun testGetRoot() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Get, "/ping").apply {
-                assertEquals(response.content, "pong")
+            handleRequest(HttpMethod.Get, "/").apply {
+                assertEquals(response.content, "Hello, world!")
                 assertEquals(HttpStatusCode.OK, response.status())
             }
         }
